fixup! JAMES-2285 Expose a method helping comparing MimeMessage as strings
